+++ ConsoleUI/InstallScreen.cs
@@ -94,6 +94,15 @@ namespace CKAN.ConsoleUI {
                         RaiseError("Game files reverted.");
                     } catch (DownloadErrorsKraken ex) {
                         RaiseError(ex.ToString());
+                    } catch (ModuleDownloadErrorsKraken ex) {
+                        RaiseError(ex.ToString());
+                    } catch (DownloadThrottledKraken ex) {
+                        if (RaiseYesNoDialog($"{ex.ToString()}\n\nEdit authentication tokens now?")) {
+                            if (ex.infoUrl != null) {
+                                ModInfoScreen.LaunchURL(ex.infoUrl);
+                            }
+                            LaunchSubScreen(new AuthTokenScreen());
+                        }
                     } catch (MissingCertificateKraken ex) {
                         RaiseError(ex.ToString());
                     } catch (InconsistentKraken ex) {